requirements.txt
python==3.12
fastapi
uvicorn
pydantic

# For JWT authentication
python-jose[cryptography]
passlib[argon2]